organize folder history
