+++ lib/src/providers/improves.dart
@@ -1,6 +1,7 @@
 import 'dart:collection';
 
 import 'package:flutter/material.dart';
+import 'package:risa2/src/api/json_models/response/improve_resp.dart';
 
 import '../api/filter_models/improve_filter.dart';
 import '../api/json_models/response/improve_list_resp.dart';
@@ -28,10 +29,12 @@ class ImproveProvider extends ChangeNotifier {
     return UnmodifiableListView(_improveList);
   }
 
-  Future<void> findImprove() async {
-    setState(ViewState.busy);
+  Future<void> findImprove({bool loading = true}) async {
+    if (loading) {
+      setState(ViewState.busy);
+    }
 
-    final filter = FilterImporve(branch: "BANJARMASIN", limit: 3);
+    final filter = FilterImporve(branch: "BANJARMASIN", limit: 10);
     var error = "";
 
     try {
@@ -57,5 +60,51 @@ class ImproveProvider extends ChangeNotifier {
     if (error.isNotEmpty) {
       return Future.error(error);
     }
+  }
+
+  // Detail Improvement ========================================
+  ViewState _detailState = ViewState.idle;
+  ViewState get detailState => _detailState;
+  void setDetailState(ViewState viewState) {
+    _detailState = viewState;
+    notifyListeners();
+  }
+
+  String _detailIDSaved = "";
+  String get detailID => _detailIDSaved;
+  void setDetailID(String id) {
+    _detailIDSaved = id;
+  }
+
+  // improve detail cache
+  ImproveDetailResponseData _improveDetail = ImproveDetailResponseData(
+      "", 0, 0, "", "", "", "", "", "", "", 0, 0, false, 0, []);
+  ImproveDetailResponseData get improveDetail {
+    return _improveDetail;
+  }
+
+  // return future true jika incremet/decrement improve berhasil
+  Future<bool> incrementImprovement(ImproveChange payload) async {
+    setState(ViewState.busy);
+    var error = "";
+
+    try {
+      final response =
+          await _improveService.changeImprove(_detailIDSaved, payload);
+      if (response.error != null) {
+        error = response.error!.message;
+      } else {
+        _improveDetail = response.data!;
+      }
+    } catch (e) {
+      error = e.toString();
+    }
+
+    if (error.isNotEmpty) {
+      return Future.error(error);
+    }
+
+    await findImprove(loading: false);
+    return true;
   }
 }
