+++ lib/src/screens/check_master/add_check_master_screen.dart
@@ -5,7 +5,7 @@ import '../../api/json_models/request/checkp_req.dart';
 import '../../config/pallatte.dart';
 import '../../models/shift_chip_models.dart';
 import '../../providers/checks_master.dart';
-import '../../shared/flushbar.dart';
+import '../../shared/func_flushbar.dart';
 import '../../shared/home_like_button.dart';
 import '../../shared/ui_helpers.dart';
 import '../../utils/enums.dart';
