+++ lib/src/screens/history/history_list_fragment.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import '../../api/json_models/response/history_list_resp.dart';
 import '../../providers/histories.dart';
 import '../../shared/empty_box.dart';
-import '../../shared/flushbar.dart';
+import '../../shared/func_flushbar.dart';
 import '../../shared/func_history_dialog.dart';
 import '../../shared/history_item_widget.dart';
 import '../../utils/enums.dart';
