+++ lib/src/screens/home/dashboard_history_fragment.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 
 import '../../providers/histories.dart';
 import '../../shared/empty_box.dart';
-import '../../shared/flushbar.dart';
+import '../../shared/func_flushbar.dart';
 import '../../shared/func_history_dialog.dart';
 import '../../shared/history_item_widget.dart';
 import '../../utils/enums.dart';
